backend/requirements.txt
# EvidenceIQ Backend Dependencies

# AWS
boto3==1.35.0
botocore==1.35.0

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# File handling
python-multipart==0.0.12

# PDF generation
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2

# CORS
starlette